file transfer application successfully implemented
very bare bones, always saves to a placeholder file called "rec.txt". for some reason the text file inputs were limiting me to 7 characters. will always overwrite whatever is in this file upon receiving a new one, so have to go into terminal to cp the file. need to work on sending the file name and saving it into this filename.